server routing restructuration
